+++ kurtosis-module/impl/prelaunch_data_generator/cl_genesis/cl_genesis_data.go
@@ -1,8 +1,6 @@
 package cl_genesis
 
 type CLGenesisData struct {
-	genesisUnixTimestamp uint64
-
 	// Path to the directory holding all the CL genesis files
 	parentDirpath string
 
@@ -11,13 +9,10 @@ type CLGenesisData struct {
 	genesisSszFilepath string
 }
 
-func newCLGenesisData(genesisUnixTimestamp uint64, parentDirpath string, configYmlFilepath string, genesisSszFilepath string) *CLGenesisData {
-	return &CLGenesisData{genesisUnixTimestamp: genesisUnixTimestamp, parentDirpath: parentDirpath, configYmlFilepath: configYmlFilepath, genesisSszFilepath: genesisSszFilepath}
+func newCLGenesisData(parentDirpath string, configYmlFilepath string, genesisSszFilepath string) *CLGenesisData {
+	return &CLGenesisData{parentDirpath: parentDirpath, configYmlFilepath: configYmlFilepath, genesisSszFilepath: genesisSszFilepath}
 }
 
-func (paths *CLGenesisData) GetGenesisUnixTimestamp() uint64 {
-	return paths.genesisUnixTimestamp
-}
 func (paths *CLGenesisData) GetParentDirpath() string {
 	return paths.parentDirpath
 }