+++ docs/changelog.md
@@ -14,6 +14,7 @@
 * Set the `--subscribe-all-subnets` flag equivalents on all Beacon nodes
 * Generate the CL genesis files AFTER the EL network is mining, so that the CL network doesn't skip any important epochs (e.g. Altair, or merge fork) which causes it to get in a stuck state
 * Removed unneeded hanging-around delay that existed in wait-for-finalization logic
+* Wait until all CL nodes are up before starting to process slots
 
 # 0.2.3
 ### Features